+++ src/main/java/com/example/demo/core/order/param/CreateOrderParam.java
@@ -12,7 +12,7 @@ public record CreateOrderParam(long memberNo,
 
     public record Product(@NotEmpty String productCode,
                           @NotEmpty String productName,
-                          @Positive long quantity,
+                          @Positive int quantity,
                           @NotNull @Positive BigDecimal productAmount) {
 
         public BigDecimal getTransactionAmount() {
